requirements.txt
langchain
langchain_community
langchain_groq
langchain_huggingface
sentence_transformers
chromadb
streamlit
python-dotenv
pandas
